changed the order of setup functions
